refactor: remove classic analyze command
